Add a large legacy demo DB dump and make several marketplace-related fixes and UI tweaks. Files added/changed: - Added agricultural-crop-management-backend/newimportdb.sql (legacy full dump / demo bootstrap for quanlymuavu) and updated root newimportdb.sql. - Added Flyway migration agricultural-crop-management-backend/src/main/resources/db/migration/V34__marketplace_cart_items_id_auto_increment.sql to enable auto-increment on marketplace_cart_items.id. - Backend: updated marketplace repository and service implementations and adjusted ErrorCode definitions. - Tests: updated MarketplaceBuyerCartIntegrationTest and MarketplaceServiceTest to reflect behavior changes. - Frontend: updated FarmStorePage.css and FarmStorePage.tsx for marketplace UI tweaks. Purpose: provide a full demo seed dump for one-run demo installs, fix marketplace cart item ID handling, and align backend, tests and frontend with the change.

+++ agricultural-crop-management-backend/newimportdb.sql
@@ -324,7 +324,7 @@ CREATE TABLE stock_movements (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
 CREATE TABLE inventory_balances (
-    id INT NOT NULL PRIMARY KEY,
+    id BIGINT NOT NULL PRIMARY KEY,
     supply_lot_id INT NOT NULL,
     warehouse_id INT NOT NULL,
     location_id INT NULL,
@@ -785,7 +785,7 @@ CREATE TABLE marketplace_product_reviews (
     order_id BIGINT NOT NULL,
     order_item_id BIGINT NULL,
     buyer_user_id BIGINT NOT NULL,
-    rating TINYINT NOT NULL,
+    rating INT NOT NULL,
     comment TEXT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NULL,
